Mise en lien avec api-users + documentation avec Swagger UI

+++ src/main/java/com/square_games/api/services/GameServiceImpl.java
@@ -1,6 +1,7 @@
 package com.square_games.api.services;
 
 import com.square_games.api.DTO.GameCreationParams;
+import com.square_games.api.clients.UserClient;
 import com.square_games.api.dao.GameDao;
 import com.square_games.api.plugins.GamePlugin;
 import fr.le_campus_numerique.square_games.engine.*;
@@ -17,14 +18,24 @@ public class GameServiceImpl implements GameService  {
 
     private final List<GamePlugin> gamePlugins;
 
-    public GameServiceImpl(GameDao gameDao, List<GamePlugin> gamePlugins) {
+    private final UserClient userClient;
+
+    public GameServiceImpl(GameDao gameDao, List<GamePlugin> gamePlugins, UserClient userClient) {
         this.gameDao = gameDao;
         this.gamePlugins = gamePlugins;
+        this.userClient = userClient;
     }
 
 
     @Override
     public Game createGame(UUID userId, GameCreationParams params) {
+
+        if (!userClient.isUserValid(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Utilisateur inconnu"
+            );
+        }
 
         GamePlugin plugin = gamePlugins.stream()
                 .filter(p -> p.getClass().getSimpleName()
@@ -54,26 +65,57 @@ public class GameServiceImpl implements GameService  {
 
     @Override
     public Game getGameById(UUID userId, String gameId) {
+        if (!userClient.isUserValid(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Utilisateur inconnu"
+            );
+        }
+
         return gameDao.findById(userId, gameId);
     }
 
     @Override
     public GameStatus getGameStatus(UUID userId, String gameId) {
+        if (!userClient.isUserValid(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Utilisateur inconnu"
+            );
+        }
         Game game = getGameById(userId, gameId);
         return game.getStatus();
     }
     @Override
     public Collection<Game> getGames(UUID userId){
+        if (!userClient.isUserValid(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Utilisateur inconnu"
+            );
+        }
         return gameDao.findAll(userId);
     }
 
     @Override
     public void deleteGameById(UUID userId, String gameId) {
+        if (!userClient.isUserValid(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Utilisateur inconnu"
+            );
+        }
         gameDao.delete(userId, gameId);
     }
 
     @Override
     public Collection<Game> getOngoingGames(UUID userId) {
+        if (!userClient.isUserValid(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Utilisateur inconnu"
+            );
+        }
         return getGames(userId)
                 .stream()
                 .filter(game -> game.getStatus() == GameStatus.ONGOING)
@@ -82,25 +124,37 @@ public class GameServiceImpl implements GameService  {
 
     @Override
     public Set<CellPosition> getAllowedMoves(UUID userId, String gameId, CellPosition position) {
+        if (!userClient.isUserValid(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Utilisateur inconnu"
+            );
+        }
         Game game = gameDao.findById(userId, gameId);
 
         if (game == null) {
             throw new IllegalArgumentException("Partie inconnue");
         }
+        GamePlugin plugin = gamePlugins.stream()
+                .filter(p -> p.getGameType()
+                        .equals(game.getFactoryId()))
+                        .findFirst()
+                .orElseThrow(() -> new IllegalArgumentException("Plugin de jeu inconnu")
+                );
 
-        Token token = game.getBoard().get(position);
-
-        if (token == null) {
-            throw new IllegalArgumentException(
-                    "Aucun jeton à cette position"
-            );
-        }
-
-        return token.getAllowedMoves();
+        return plugin.getAllowedMoves(game, position);
     }
 
     @Override
     public void playMove(UUID userId, String gameId, CellPosition tokenPosition, CellPosition targetPosition) {
+
+        if (!userClient.isUserValid(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Utilisateur inconnu"
+            );
+        }
+
         Game game = gameDao.findById(userId, gameId);
 
         if (game == null) {
@@ -114,13 +168,14 @@ public class GameServiceImpl implements GameService  {
             );
         }
 
-        Token token = game.getBoard().get(tokenPosition);
+        GamePlugin plugin = gamePlugins.stream()
+                .filter(p -> p.getGameType()
+                        .equals(game.getFactoryId()))
+                .findFirst()
+                .orElseThrow(() -> new IllegalArgumentException("Plugin de jeu inconnu")
+                );
 
-        if (token == null) {
-            throw new IllegalArgumentException(
-                    "Aucun jeton à cette position"
-            );
-        }
+        Token token = plugin.getTokenToMove(game, tokenPosition);
 
         try {
             token.moveTo(targetPosition);
